Camada service finalizada <3

+++ src/main/java/com/example/VOM_HiveJava/dto/request/AbstractRequest.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.NotNull;
 
 public record AbstractRequest(
         @NotNull(message = "O id é campo obrigatório")
-        int id
+        Long id
 ) {
 }
