+++ src/main/java/com/example/VOM_HiveJava/entity/Pay.java
@@ -21,7 +21,7 @@ public class Pay {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "SQ_PAY_HIST")
     @SequenceGenerator(name = "SQ_PAY_HIST", sequenceName = "SQ_PAY_HIST", allocationSize = 1)
     @Column(name = "ID_PAY_HIST")
-    private int id_history;
+    private Long id_history;
 
     @Column(name = "VALUE_HISTORY")
     private double value;
